Extending python tool to show more command-line arguments to control which python is used and where py/pyext files should be installed
The new options are in addition to the existing PYTHON, PYTHONDIR, and
PYTHONARCHDIR, providing help messages to the user, so it can easily
change parameters without second-guessing of what are they.

Environmental parameters override command line arguments (is it right?)

Also, this commit includes extension to Configure.find_command that
allows to specify name of the searched command. This is helpful when
search command is a huge list, as in case of python's python-config case.