Correction of boost module detection of Boost.Log
The previous logic made an incorrect assumption that the filename of
shared/static library indicates that it was build as multi-threaded or
single threaded.  This assumption does not hold in many Linux
distributions.

In addition to that. Boost.Thread and Boost.Log require -pthread (or
some other) flags in order to properly link.